Folder/File structure change to Vitruvius Kinect v1 to match structure of Vitruvius Kinect v2 more closely: - moved GestureController.cs to folder Gestures - merged all SwipeDownSegment?.cs files (where ?=1,2,3) into a single SwipeDownSegments.cs file - merged all SwipeUpSegment?.cs (where ?=1,2,3) into a single SwipeUpSegments.cs file - moved all *Segments.cs files into a new Gestures/Segments subfolder, instead of keeping them in separate subfolders - Copied Namespace.doc file from Vitruvius Kinect v2 WPF project into Gestures subfolder
