+++ Kinect v1/LightBuzz.Vitruvius/Gestures/GestureController.cs
