+++ Kinect v1/LightBuzz.Vitruvius/Gestures/Segments/JoinedHandsSegment.cs
