+++ Kinect v1/LightBuzz.Vitruvius/Gestures/Segments/MenuSegments.cs
