+++ Kinect v1/LightBuzz.Vitruvius/Gestures/Segments/SwipeLeftSegments.cs
