+++ Kinect v1/LightBuzz.Vitruvius/Gestures/Segments/SwipeRightSegments.cs
