+++ Kinect v1/LightBuzz.Vitruvius/Gestures/Segments/SwipeUpSegments.cs
@@ -1,0 +1,104 @@
+﻿using Microsoft.Kinect;
+using System.Diagnostics;
+
+namespace LightBuzz.Vitruvius.Gestures
+{
+    /// <summary>
+    /// The first part of the swipe up gesture
+    /// </summary>
+    public class SwipeUpSegment1 : IGestureSegment
+    {
+        /// <summary>
+        /// Updates the current gesture.
+        /// </summary>
+        /// <param name="skeleton">The skeleton.</param>
+        /// <returns>A GesturePartResult based on whether the gesture part has been completed.</returns>
+        public GesturePartResult Update(Skeleton skeleton)
+        {
+
+            // right hand in front of right elbow
+            if (skeleton.Joints[JointType.HandRight].Position.Z < skeleton.Joints[JointType.ElbowRight].Position.Z)
+            {
+                // right hand below shoulder height but above hip height
+                if (skeleton.Joints[JointType.HandRight].Position.Y < skeleton.Joints[JointType.Head].Position.Y && skeleton.Joints[JointType.HandRight].Position.Y > skeleton.Joints[JointType.HipCenter].Position.Y)
+                {
+                    // right hand right of right shoulder
+                    if (skeleton.Joints[JointType.HandRight].Position.X > skeleton.Joints[JointType.ShoulderRight].Position.X)
+                    {
+                        return GesturePartResult.Succeeded;
+                    }
+                    return GesturePartResult.Undetermined;
+                }
+                return GesturePartResult.Failed;
+            }
+            return GesturePartResult.Failed;
+        }
+    }
+
+    /// <summary>
+    /// The second part of the swipe up gesture
+    /// </summary>
+    public class SwipeUpSegment2 : IGestureSegment
+    {
+        /// <summary>
+        /// Updates the current gesture.
+        /// </summary>
+        /// <param name="skeleton">The skeleton.</param>
+        /// <returns>A GesturePartResult based on whether the gesture part has been completed.</returns>
+        public GesturePartResult Update(Skeleton skeleton)
+        {
+            // right hand in front of right shoulder
+            if (skeleton.Joints[JointType.HandRight].Position.Z < skeleton.Joints[JointType.ShoulderRight].Position.Z)
+            {
+                // right hand above right shoulder
+                if (skeleton.Joints[JointType.HandRight].Position.Y > skeleton.Joints[JointType.ShoulderRight].Position.Y)
+                {
+                    // right hand right of right shoulder
+                    if (skeleton.Joints[JointType.HandRight].Position.X > skeleton.Joints[JointType.ShoulderRight].Position.X)
+                    {
+                        return GesturePartResult.Succeeded;
+                    }
+                    return GesturePartResult.Undetermined;
+                }
+                return GesturePartResult.Failed;
+            }
+            return GesturePartResult.Failed;
+        }
+    }
+    
+    /// <summary>
+    /// The third part of the swipe up gesture
+    /// </summary>
+    public class SwipeUpSegment3 : IGestureSegment
+    {
+        /// <summary>
+        /// Updates the current gesture.
+        /// </summary>
+        /// <param name="skeleton">The skeleton.</param>
+        /// <returns>A GesturePartResult based on whether the gesture part has been completed.</returns>
+        public GesturePartResult Update(Skeleton skeleton)
+        {
+            // //Right hand in front of right shoulder
+            if (skeleton.Joints[JointType.HandRight].Position.Z < skeleton.Joints[JointType.ShoulderRight].Position.Z)
+            {
+                // right hand above head
+                if (skeleton.Joints[JointType.HandRight].Position.Y > skeleton.Joints[JointType.Head].Position.Y)
+                {
+                    // right hand right of right shoulder
+                    if (skeleton.Joints[JointType.HandRight].Position.X > skeleton.Joints[JointType.ShoulderRight].Position.X)
+                    {
+                        return GesturePartResult.Succeeded;
+                    }
+                    return GesturePartResult.Undetermined;
+                }
+
+                // Debug.WriteLine("GesturePart 2 - right hand below shoulder height but above hip height - FAIL");
+                return GesturePartResult.Failed;
+            }
+
+            // Debug.WriteLine("GesturePart 2 - Right hand in front of right Shoulder - FAIL");
+            return GesturePartResult.Failed;
+        }
+    }
+
+}
