+++ Kinect v1/LightBuzz.Vitruvius/Gestures/Segments/WaveLeftSegments.cs
