+++ Kinect v1/LightBuzz.Vitruvius/Gestures/Segments/WaveRightSegments.cs
