+++ Kinect v1/LightBuzz.Vitruvius/Gestures/Segments/ZoomSegments.cs
