--- Kinect v1/LightBuzz.Vitruvius/Gestures/SwipeDown/SwipeDownSegment2.cs
@@ -1,35 +1,0 @@
-﻿using Microsoft.Kinect;
-
-namespace LightBuzz.Vitruvius.Gestures
-{
-    /// <summary>
-    /// The second part of the swipe down gesture for the right hand
-    /// </summary>
-    public class SwipeDownSegment2 : IGestureSegment
-    {
-        /// <summary>
-        /// Updates the current gesture.
-        /// </summary>
-        /// <param name="skeleton">The skeleton.</param>
-        /// <returns>A GesturePartResult based on whether the gesture part has been completed.</returns>
-        public GesturePartResult Update(Skeleton skeleton)
-        {
-            // right hand in front of right shoulder
-            if (skeleton.Joints[JointType.HandRight].Position.Z < skeleton.Joints[JointType.ElbowRight].Position.Z && skeleton.Joints[JointType.HandRight].Position.Y < skeleton.Joints[JointType.ShoulderCenter].Position.Y)
-            {
-                // right hand below right elbow
-                if (skeleton.Joints[JointType.HandRight].Position.Y < skeleton.Joints[JointType.ElbowRight].Position.Y)
-                {
-                    // right hand right of right shoulder
-                    if (skeleton.Joints[JointType.HandRight].Position.X > skeleton.Joints[JointType.HipRight].Position.X)
-                    {
-                        return GesturePartResult.Succeeded;
-                    }
-                    return GesturePartResult.Undetermined;
-                }
-                return GesturePartResult.Failed;
-            }
-            return GesturePartResult.Failed;
-        }
-    }
-}
