--- Kinect v1/LightBuzz.Vitruvius/Gestures/SwipeUp/SwipeUpSegment2.cs
@@ -1,35 +1,0 @@
-﻿using Microsoft.Kinect;
-
-namespace LightBuzz.Vitruvius.Gestures
-{
-    /// <summary>
-    /// The second part of the swipe up gesture
-    /// </summary>
-    public class SwipeUpSegment2 : IGestureSegment
-    {
-        /// <summary>
-        /// Updates the current gesture.
-        /// </summary>
-        /// <param name="skeleton">The skeleton.</param>
-        /// <returns>A GesturePartResult based on whether the gesture part has been completed.</returns>
-        public GesturePartResult Update(Skeleton skeleton)
-        {
-            // right hand in front of right shoulder
-            if (skeleton.Joints[JointType.HandRight].Position.Z < skeleton.Joints[JointType.ShoulderRight].Position.Z)
-            {
-                // right hand above right shoulder
-                if (skeleton.Joints[JointType.HandRight].Position.Y > skeleton.Joints[JointType.ShoulderRight].Position.Y)
-                {
-                    // right hand right of right shoulder
-                    if (skeleton.Joints[JointType.HandRight].Position.X > skeleton.Joints[JointType.ShoulderRight].Position.X)
-                    {
-                        return GesturePartResult.Succeeded;
-                    }
-                    return GesturePartResult.Undetermined;
-                }
-                return GesturePartResult.Failed;
-            }
-            return GesturePartResult.Failed;
-        }
-    }
-}
