--- Kinect v1/LightBuzz.Vitruvius/Gestures/SwipeUp/SwipeUpSegment3.cs
@@ -1,39 +1,0 @@
-﻿using Microsoft.Kinect;
-
-namespace LightBuzz.Vitruvius.Gestures
-{
-    /// <summary>
-    /// The third part of the swipe up gesture
-    /// </summary>
-    public class SwipeUpSegment3 : IGestureSegment
-    {
-        /// <summary>
-        /// Updates the current gesture.
-        /// </summary>
-        /// <param name="skeleton">The skeleton.</param>
-        /// <returns>A GesturePartResult based on whether the gesture part has been completed.</returns>
-        public GesturePartResult Update(Skeleton skeleton)
-        {
-            // //Right hand in front of right shoulder
-            if (skeleton.Joints[JointType.HandRight].Position.Z < skeleton.Joints[JointType.ShoulderRight].Position.Z)
-            {
-                // right hand above head
-                if (skeleton.Joints[JointType.HandRight].Position.Y > skeleton.Joints[JointType.Head].Position.Y)
-                {
-                    // right hand right of right shoulder
-                    if (skeleton.Joints[JointType.HandRight].Position.X > skeleton.Joints[JointType.ShoulderRight].Position.X)
-                    {
-                        return GesturePartResult.Succeeded;
-                    }
-                    return GesturePartResult.Undetermined;
-                }
-
-                // Debug.WriteLine("GesturePart 2 - right hand below shoulder height but above hip height - FAIL");
-                return GesturePartResult.Failed;
-            }
-
-            // Debug.WriteLine("GesturePart 2 - Right hand in front of right Shoulder - FAIL");
-            return GesturePartResult.Failed;
-        }
-    }
-}
