+++ Kinect v1/LightBuzz.Vitruvius/LightBuzz.Vitruvius.csproj
@@ -46,25 +46,22 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Gestures\Gesture.cs" />
-    <Compile Include="GestureController.cs" />
+    <Compile Include="Gestures\GestureController.cs" />
     <Compile Include="Gestures\GesturePartResult.cs" />
     <Compile Include="Gestures\GestureEventArgs.cs" />
     <Compile Include="Gestures\GestureType.cs" />
     <Compile Include="Gestures\IGestureSegment.cs" />
+    <Compile Include="Gestures\NamespaceDoc.cs" />
+    <Compile Include="Gestures\Segments\JoinedHandsSegment.cs" />
+    <Compile Include="Gestures\Segments\MenuSegments.cs" />
+    <Compile Include="Gestures\Segments\SwipeDownSegments.cs" />
+    <Compile Include="Gestures\Segments\SwipeLeftSegments.cs" />
+    <Compile Include="Gestures\Segments\SwipeRightSegments.cs" />
+    <Compile Include="Gestures\Segments\SwipeUpSegments.cs" />
+    <Compile Include="Gestures\Segments\WaveLeftSegments.cs" />
+    <Compile Include="Gestures\Segments\WaveRightSegments.cs" />
+    <Compile Include="Gestures\Segments\ZoomSegments.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Gestures\JoinedHands\JoinedHandsSegment.cs" />
-    <Compile Include="Gestures\Menu\MenuSegments.cs" />
-    <Compile Include="Gestures\SwipeDown\SwipeDownSegment1.cs" />
-    <Compile Include="Gestures\SwipeDown\SwipeDownSegment2.cs" />
-    <Compile Include="Gestures\SwipeDown\SwipeDownSegment3.cs" />
-    <Compile Include="Gestures\SwipeLeft\SwipeLeftSegments.cs" />
-    <Compile Include="Gestures\SwipeRight\SwipeRightSegments.cs" />
-    <Compile Include="Gestures\SwipeUp\SwipeUpSegment1.cs" />
-    <Compile Include="Gestures\SwipeUp\SwipeUpSegment2.cs" />
-    <Compile Include="Gestures\SwipeUp\SwipeUpSegment3.cs" />
-    <Compile Include="Gestures\WaveLeft\WaveLeftSegments.cs" />
-    <Compile Include="Gestures\WaveRight\WaveRightSegments.cs" />
-    <Compile Include="Gestures\Zoom\ZoomSegments.cs" />
     <Compile Include="SensorExtensions.cs" />
     <Compile Include="SkeletonExtensions.cs" />
     <Compile Include="VisualizationMode.cs" />
